[NN CZ] unit tests for orbit model package server to use job id

+++ foundations_model_package/src/integration/fixtures/model-server-with-recalibrate/project_code/driver.py
@@ -1,4 +1,6 @@
 import pickle
+import os
 
-with open('model_file', 'wb') as model_file:
-    pickle.dump(7, model_file)
+if not os.path.isfile('model_file'):
+    with open('model_file', 'wb') as model_file:
+        pickle.dump(7, model_file)
